requirements.txt
google-generativeai
flask
flask-httpauth
requests
werkzeug